+++ Lab2/app/get_way.py
@@ -4,18 +4,13 @@ def json_unpack() -> list:
     """
     Open jsons files
     """
-    path_to_dir = os.path.dirname(__file__)
 
-    with open(os.path.join(path_to_dir, "json", "src.json"), 'r') as src_json:
-        src = json.load(src_json)
-
-    with open(os.path.join(path_to_dir, "json", "setting.json"), 'r') as setting_json:
+    with open(os.path.join(os.path.dirname(__file__), "json", "setting.json"), 'r') as setting_json:
         setting = json.load(setting_json)
 
-    with open(os.path.join(path_to_dir, "json", "name.json"), 'r') as name_json:
-        name = json.load(name_json)
-
-    return [os.path.join(src["name-images-dir"]), os.path.join(src["name-dir"], src["name-data-dir"], name["name-together"]), os.path.join(src["name-dir"], src["name-data-dir"], name["name-random"])]
+    return [setting["name-images-dir"], 
+            os.path.join(setting["name-dir"], setting["name-data-dir"], setting["name-together"]), 
+            os.path.join(setting["name-dir"], setting["name-data-dir"], setting["name-random"])]
 
 
 def get_path_normal(object : str, number : int) -> str:
@@ -29,7 +24,7 @@ def get_path_normal(object : str, number : int) -> str:
     number : int
         Number of object
     """
-    return os.path.join(json_unpack()[0], f"{object}/{number:04}.jpg")
+    return os.path.join(json_unpack()[0], object, f"{number:04}.jpg")
 
 
 def get_path_together(object : str, number : int) -> str:
