+++ Lab2/app/json/setting.json
@@ -1,0 +1,14 @@
+{
+    "mode" : "normal",
+    "objects" : ["rose", "tulip"],
+
+    "name-normal" : "dataset", 
+    "name-together" : "dataset_together", 
+    "name-random" : "dataset_random",
+
+    "name-dir" : "app",
+    "name-csv-dir" : "csv",
+    "name-data-dir" : "data",
+    "name-json-dir" : "json",
+    "name-images-dir" : "../Lab1/dataset"
+}
