--- app/json/name.json
@@ -1,5 +1,0 @@
-{
-    "name-normal" : "dataset", 
-    "name-together" : "dataset_together", 
-    "name-random" : "dataset_random"
-}
