--- app/json/setting.json
@@ -1,4 +1,0 @@
-{
-    "mode" : "random",
-    "objects" : ["rose"]
-}
