--- app/json/src.json
@@ -1,7 +1,0 @@
-{
-    "name-dir" : "app",
-    "name-csv-dir" : "csv",
-    "name-data-dir" : "data",
-    "name-json-dir" : "json",
-    "name-images-dir" : "Lab1/dataset"
-}
